Refactored provider classes to provider.ts
