+++ Extension/src/LanguageServer/client.ts
@@ -18,6 +18,7 @@ import { CppSettings, OtherSettings } from './settings';
 import * as telemetry from '../telemetry';
 import { PersistentState, PersistentFolderState } from './persistentState';
 import { UI, getUI } from './ui';
+import { FoldingRangeProvider, SemanticTokensProvider } from './providers';
 import { ClientCollection } from './clientCollection';
 import { createProtocolFilter } from './protocolFilter';
 import { DataBinding } from './dataBinding';
@@ -55,7 +56,7 @@ let debugChannel: vscode.OutputChannel;
 let diagnosticsCollection: vscode.DiagnosticCollection;
 let workspaceDisposables: vscode.Disposable[] = [];
 let workspaceReferences: refs.ReferencesManager;
-const openFileVersions: Map<string, number> = new Map<string, number>();
+export const openFileVersions: Map<string, number> = new Map<string, number>();
 
 export function disposeWorkspaceData(): void {
     workspaceDisposables.forEach((d) => d.dispose());
@@ -297,12 +298,12 @@ interface DidChangeConfigurationParams extends WorkspaceFolderParams {
     settings: any;
 }
 
-interface GetFoldingRangesParams {
+export interface GetFoldingRangesParams {
     uri: string;
     id: number;
 }
 
-enum FoldingRangeKind {
+export enum FoldingRangeKind {
     None = 0,
     Comment = 1,
     Imports = 2,
@@ -323,7 +324,7 @@ interface AbortRequestParams {
     id: number;
 }
 
-interface GetSemanticTokensParams {
+export interface GetSemanticTokensParams {
     uri: string;
     id: number;
 }
@@ -387,8 +388,8 @@ const GetDiagnosticsRequest: RequestType<void, GetDiagnosticsResult, void, void>
 const GetCodeActionsRequest: RequestType<GetCodeActionsRequestParams, CodeActionCommand[], void, void> = new RequestType<GetCodeActionsRequestParams, CodeActionCommand[], void, void>('cpptools/getCodeActions');
 const GetDocumentSymbolRequest: RequestType<GetDocumentSymbolRequestParams, LocalizeDocumentSymbol[], void, void> = new RequestType<GetDocumentSymbolRequestParams, LocalizeDocumentSymbol[], void, void>('cpptools/getDocumentSymbols');
 const GetSymbolInfoRequest: RequestType<WorkspaceSymbolParams, LocalizeSymbolInformation[], void, void> = new RequestType<WorkspaceSymbolParams, LocalizeSymbolInformation[], void, void>('cpptools/getWorkspaceSymbols');
-const GetFoldingRangesRequest: RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void> = new RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void>('cpptools/getFoldingRanges');
-const GetSemanticTokensRequest: RequestType<GetSemanticTokensParams, GetSemanticTokensResult, void, void> = new RequestType<GetSemanticTokensParams, GetSemanticTokensResult, void, void>('cpptools/getSemanticTokens');
+export const GetFoldingRangesRequest: RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void> = new RequestType<GetFoldingRangesParams, GetFoldingRangesResult, void, void>('cpptools/getFoldingRanges');
+export const GetSemanticTokensRequest: RequestType<GetSemanticTokensParams, GetSemanticTokensResult, void, void> = new RequestType<GetSemanticTokensParams, GetSemanticTokensResult, void, void>('cpptools/getSemanticTokens');
 
 // Notifications to the server
 const DidOpenNotification: NotificationType<DidOpenTextDocumentParams, void> = new NotificationType<DidOpenTextDocumentParams, void>('textDocument/didOpen');
@@ -448,7 +449,6 @@ interface ReferencesCancellationState {
 
 const referencesPendingCancellations: ReferencesCancellationState[] = [];
 
-let abortRequestId: number = 0;
 
 class ClientModel {
     public isTagParsing: DataBinding<boolean>;
@@ -553,109 +553,6 @@ export function createNullClient(): Client {
     return new NullClient();
 }
 
-class FoldingRangeProvider implements vscode.FoldingRangeProvider {
-    private client: DefaultClient;
-    constructor(client: DefaultClient) {
-        this.client = client;
-    }
-    provideFoldingRanges(document: vscode.TextDocument, context: vscode.FoldingContext,
-        token: vscode.CancellationToken): Promise<vscode.FoldingRange[]> {
-        const id: number = ++abortRequestId;
-        const params: GetFoldingRangesParams = {
-            id: id,
-            uri: document.uri.toString()
-        };
-        return new Promise<vscode.FoldingRange[]>((resolve, reject) => {
-            this.client.notifyWhenReady(() => {
-                this.client.languageClient.sendRequest(GetFoldingRangesRequest, params)
-                    .then((ranges) => {
-                        if (ranges.canceled) {
-                            reject();
-                        } else {
-                            const result: vscode.FoldingRange[] = [];
-                            ranges.ranges.forEach((r) => {
-                                const foldingRange: vscode.FoldingRange = {
-                                    start: r.range.start.line,
-                                    end: r.range.end.line
-                                };
-                                switch (r.kind) {
-                                    case FoldingRangeKind.Comment:
-                                        foldingRange.kind = vscode.FoldingRangeKind.Comment;
-                                        break;
-                                    case FoldingRangeKind.Imports:
-                                        foldingRange.kind = vscode.FoldingRangeKind.Imports;
-                                        break;
-                                    case FoldingRangeKind.Region:
-                                        foldingRange.kind = vscode.FoldingRangeKind.Region;
-                                        break;
-                                    default:
-                                        break;
-                                }
-                                result.push(foldingRange);
-                            });
-                            resolve(result);
-                        }
-                    });
-                token.onCancellationRequested(e => this.client.abortRequest(id));
-            });
-        });
-    }
-}
-
-class SemanticTokensProvider implements vscode.DocumentSemanticTokensProvider {
-    private client: DefaultClient;
-    public onDidChangeSemanticTokensEvent = new vscode.EventEmitter<void>();
-    public onDidChangeSemanticTokens?: vscode.Event<void>;
-    private tokenCaches: Map<string, [number, vscode.SemanticTokens]> = new Map<string, [number, vscode.SemanticTokens]>();
-
-    constructor(client: DefaultClient) {
-        this.client = client;
-        this.onDidChangeSemanticTokens = this.onDidChangeSemanticTokensEvent.event;
-    }
-
-    public async provideDocumentSemanticTokens(document: vscode.TextDocument, token: vscode.CancellationToken): Promise<vscode.SemanticTokens> {
-        return new Promise<vscode.SemanticTokens>((resolve, reject) => {
-            this.client.notifyWhenReady(() => {
-                const uriString: string = document.uri.toString();
-                // First check the token cache to see if we already have results for that file and version
-                const cache: [number, vscode.SemanticTokens] | undefined = this.tokenCaches.get(uriString);
-                if (cache && cache[0] === document.version) {
-                    resolve(cache[1]);
-                } else {
-                    const id: number = ++abortRequestId;
-                    const params: GetSemanticTokensParams = {
-                        id: id,
-                        uri: uriString
-                    };
-                    this.client.languageClient.sendRequest(GetSemanticTokensRequest, params)
-                        .then((tokensResult) => {
-                            if (tokensResult.canceled) {
-                                reject();
-                            } else {
-                                if (tokensResult.fileVersion !== openFileVersions.get(uriString)) {
-                                    reject();
-                                } else {
-                                    const builder: vscode.SemanticTokensBuilder = new vscode.SemanticTokensBuilder(this.client.semanticTokensLegend);
-                                    tokensResult.tokens.forEach((token) => {
-                                        builder.push(token.line, token.character, token.length, token.type, token.modifiers);
-                                    });
-                                    const tokens: vscode.SemanticTokens = builder.build();
-                                    this.tokenCaches.set(uriString, [tokensResult.fileVersion, tokens]);
-                                    resolve(tokens);
-                                }
-                            }
-                        });
-                    token.onCancellationRequested(e => this.client.abortRequest(id));
-                }
-            });
-        });
-    }
-
-    public invalidateFile(uri: string): void {
-        this.tokenCaches.delete(uri);
-        this.onDidChangeSemanticTokensEvent.fire();
-    }
-}
 
 export class DefaultClient implements Client {
     private innerLanguageClient?: LanguageClient; // The "client" that launches and communicates with our language "server" process.
